backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.53
google-generativeai==0.3.2
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
wikipedia==1.4.0
python-dotenv==1.0.0
duckduckgo-search==4.1.1
pytest==7.4.3
pytest-asyncio==0.21.1